updates and fixes moving away from separate servers move code into unified folders
