+++ config/default.json
@@ -1,6 +1,5 @@
 {
-  "PORT": 4000,
-  "ADMIN_PORT": 3001,
+  "PORT": 3001,
   "HOSTNAME": "http://central-ledger",
   "RUN_MIGRATIONS": true,
   "AMOUNT": {
@@ -8,7 +7,7 @@
     "SCALE": 2
   },
   "SIDECAR": {
-    "DISABLED": false,
+    "DISABLED": true,
     "HOST": "localhost",
     "PORT": 5678,
     "CONNECT_TIMEOUT": 45000,
