+++ package.json
@@ -20,7 +20,7 @@
     ]
   },
   "scripts": {
-    "start": "run-p start:api start:admin",
+    "start": "CLEDG_HOSTNAME=http://localhost:3001 node src/endpoints/index.js",
     "start:api": "CLEDG_HOSTNAME=http://localhost:3000 node src/api/index.js",
     "start:handlers": "node src/handlers/index.js",
     "start:admin": "CLEDG_ADMIN_KEY=admin CLEDG_ADMIN_SECRET=admin CLEDG_HOSTNAME=http://localhost:3001 node src/admin/index.js",
