+++ src/endpoints/metrics/handler.js
