+++ src/endpoints/metrics/plugin.js
