+++ src/endpoints/metrics/routes.js
