+++ src/endpoints/routes.js
@@ -3,9 +3,9 @@
 const Glob = require('glob')
 
 exports.plugin = {
-  name: 'api routes',
+  name: 'routes',
   register: function (server, options) {
-    Glob.sync('**/routes.js', {cwd: __dirname, ignore: 'routes.js'})
+    Glob.sync('**/routes.js', { cwd: __dirname, ignore: 'routes.js' })
       .forEach(x => server.route(require('./' + x)))
   }
 }
