+++ src/lib/config.js
@@ -3,7 +3,7 @@ const RC = require('rc')('CLEDG', require('../../config/default.json'))
 module.exports = {
   HOSTNAME: RC.HOSTNAME.replace(/\/$/, ''),
   PORT: RC.PORT,
-  ADMIN_PORT: RC.ADMIN_PORT,
+  // ADMIN_PORT: RC.ADMIN_PORT,
   DATABASE_URI: RC.DATABASE_URI,
   AMOUNT: RC.AMOUNT,
   EXPIRES_TIMEOUT: RC.EXPIRES_TIMEOUT,
