+++ test/unit/endpoints/routes.test.js
@@ -25,37 +25,12 @@
 'use strict'
 
 const Test = require('tape')
-const Base = require('../../base')
-const AdminRoutes = require('../../../../src/admin/routes')
-const Sinon = require('sinon')
-const Enum = require('../../../../src/lib/enum')
-const P = require('bluebird')
+const Base = require('../base')
+const ApiRoutes = require('../../../src/endpoints/routes')
 
-Test('test root routes - health', async function (assert) {
+Test('test health', async function (assert) {
   let req = Base.buildRequest({ url: '/health', method: 'GET' })
-  const server = await Base.setup(AdminRoutes)
-  const res = await server.inject(req)
-  assert.ok(res)
-  await server.stop()
-  assert.end()
-})
-
-Test('test root routes - enums', async function (assert) {
-  let sandbox = Sinon.createSandbox()
-
-  sandbox.stub(Enum, 'all').returns(P.resolve({}))
-  let req = Base.buildRequest({ url: '/enums', method: 'GET' })
-  const server = await Base.setup(AdminRoutes)
-  const res = await server.inject(req)
-  assert.ok(res)
-  sandbox.restore()
-  await server.stop()
-  assert.end()
-})
-
-Test('test root routes - /', async function (assert) {
-  let req = Base.buildRequest({ url: '/', method: 'GET' })
-  const server = await Base.setup(AdminRoutes)
+  const server = await Base.setup(ApiRoutes)
   const res = await server.inject(req)
   assert.ok(res)
   await server.stop()
